+++ projects/mainio-forms/src/lib/interfaces/i-question-display-validator.ts
@@ -1,0 +1,12 @@
+import { QuestionBase } from "../models/question-base";
+
+export abstract class IQuestionDisplayValidator {
+  abstract forForm: string;
+
+  abstract validate(
+    formId: string,
+    validatingQuestion: QuestionBase<any>,
+    currentValue: boolean,
+    questions: QuestionBase<any>[]
+  );
+}