+++ projects/mainio-forms/src/lib/models/form.ts
@@ -1,11 +1,12 @@
 import { QuestionBase } from "./question-base";
 import { FormGroup } from "@angular/forms";
 import { QuestionGroup } from "./question-group";
+import { IQuestionBaseOptions } from "../interfaces";
 
 export class Form {
   public id: string;
   public questionGroups?: { key?: string; value?: QuestionGroup };
-  public questions?: QuestionBase<any>[];
+  public questions?: IQuestionBaseOptions[];
   public values?: { key: string; value: any };
   public savedValues?: { key: string; value: any };
   public mappedModel?: any;