+++ projects/mainio-forms/src/lib/services/index.ts
@@ -5,3 +5,4 @@ export * from "./question-group.service";
 export * from "./question-json-parser.service";
 export * from "./form-data-mapper.service";
 export * from "./question-component-factory.service";
+export * from "./question-display-validator.service";